refactor: consolidate agent runtime primitives
